Adding controllers and add update request in userController + some validations rules for inputs formats
