+++ api/controllers/users.js
@@ -1,13 +1,12 @@
+const mongoose = require('mongoose');
 const express = require('express');
 const router = express.Router();
-const mongoose = require('mongoose');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-
 const User = require('../models/user');
 
-// Create new user
-router.post('/signup', (req, res, next) => {
+
+exports.user_signup = (req, res, next) => {
     User.find({email: req.body.email})
     .exec()
     .then(user => {
@@ -46,68 +45,9 @@ router.post('/signup', (req, res, next) => {
             });
         }
     });
-});
+};
 
-// Get all Users
-router.get('/', (req, res, next) => {
-    User.find()
-    .select('name email')
-    .exec()
-    .then(docs => {
-        const response = {
-            count: docs.length,
-            users: docs.map(doc => {
-                return {
-                    _id: doc._id,
-                    name: doc.name,
-                    email: doc.email,
-                    url: {
-                         type: 'GET',
-                         url: 'http://localhost:3000/user/' + doc._id //nom du domaine
-                    }
-                }
-            })
-        }
-        res.status(200).json(response)
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({
-            error: err
-        })
-    });
- });
-
-// Get user by ID
-router.get('/:userId', (req, res, next) => {
-    const id = req.params.userId;
-    User.findById(id)
-    .select('name email')
-    .exec()
-    .then(doc => {
-        console.log('From database', doc);
-        if (doc) {
-            res.status(200).json({
-                user: doc,
-                request: {
-                    type: 'GET',
-                    description: 'GET_USER_BY_ID',
-                    url: 'http://localhost:3000/user/' + doc._id 
-
-                }
-            });
-        } else {
-            res.status(404).json({message: 'No valid entry found for provided ID'});
-        }
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({error: err});
-    });
-});
-
-// Login
-router.post('/login', (req, res, next) => {
+exports.users_login = (req, res, next) => {
     User.find({ email: req.body.email})
     .exec()
     .then(user => {
@@ -149,10 +89,89 @@ router.post('/login', (req, res, next) => {
             error: err
         });
     });
-})
+};
 
-//Delete user by Id
-router.delete('/:userId', (req, res, next) => {
+exports.users_get_all = (req, res, next) => {
+    User.find()
+    .select('name email')
+    .exec()
+    .then(docs => {
+        const response = {
+            count: docs.length,
+            users: docs.map(doc => {
+                return {
+                    _id: doc._id,
+                    name: doc.name,
+                    email: doc.email,
+                    url: {
+                         type: 'GET',
+                         url: 'http://localhost:3000/user/' + doc._id //nom du domaine
+                    }
+                }
+            })
+        }
+        res.status(200).json(response)
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        })
+    });
+ };
+
+ exports.users_get_user = (req, res, next) => {
+    const id = req.params.userId;
+    User.findById(id)
+    .select('name email')
+    .exec()
+    .then(doc => {
+        console.log('From database', doc);
+        if (doc) {
+            res.status(200).json({
+                user: doc,
+                request: {
+                    type: 'GET',
+                    description: 'GET_USER_BY_ID',
+                    url: 'http://localhost:3000/user/' + doc._id 
+
+                }
+            });
+        } else {
+            res.status(404).json({message: 'No valid entry found for provided ID'});
+        }
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({error: err});
+    });
+};
+exports.users_update_user = (req, res, next) => {
+    const id = req.params.userId;
+    const updateOps = {};
+    for (const ops of req.body) {
+        updateOps[ops.propUser] = ops.value;
+    }
+    User.update({ _id: id }, {$set: updateOps})
+    .exec()
+    .then(result => {
+        console.log(result);
+        res.status(200).json({
+            message:'User UPDATED successfully !',
+            request: {
+                type: 'GET',
+                description: 'GET_User_BY_ID',
+                url: 'http://localhost:3000/users/' + id 
+            },
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({error: err});
+    });
+};
+
+exports.users_delete_user =  (req, res, next) => {
     User.remove({_id: req.params.userId})
     .exec()
     .then(result => {
@@ -166,5 +185,4 @@ router.delete('/:userId', (req, res, next) => {
             error: err
         });
     });
-})
-module.exports = router;
+};
