+++ api/models/group.js
@@ -4,9 +4,9 @@ const groupSchema = mongoose.Schema({
     _id: mongoose.Schema.Types.ObjectId,
     name: {
         type: String,
-        required: true,
-        min: 3,
-        max: 20
+        required: [true, , 'Groupe name is required'],
+        match: /[a-z]/,
+        minlength: [3, 'Group name too long !'],
     },
     user: {
         type: String,
